Add Python setup and basics tasks

+++ Python Setup & Basics/Task_3.py
@@ -1,4 +1,5 @@
-celsius = float(input("Enter Celsius: "))
+#write a program to convert celsius to fahrenhit
+celsius = int(input("Enter Celsius: "))
 
 fahrenheit = (celsius * 9 / 5) + 32
